[#25] Move adapter to his package
