+++ app/src/main/java/es/craftsmanship/toledo/katangapp/adapters/BusStopsAdapter.java
@@ -1,7 +1,6 @@
 package es.craftsmanship.toledo.katangapp.adapters;
 
 import es.craftsmanship.toledo.katangapp.activities.BusStopActivity;
-import es.craftsmanship.toledo.katangapp.activities.BusStopRoutesAdapter;
 import es.craftsmanship.toledo.katangapp.activities.R;
 import es.craftsmanship.toledo.katangapp.models.BusStop;
 import es.craftsmanship.toledo.katangapp.models.BusStopResult;
